9 screenshots of concoctions made via replit.com
I started programming freshman year of high school through college to now and I only JUST found out about this "replit.com" - it's basically an Integrated Development Environment (IDE) like visual studio but accessed and operated through your browser. This is mad clutch because you don't have to download and install an entire IDE (or at least the package add-ons for a language you're just now starting to use).

You just go to the site (https://replit.com/) and pick a language and then it presents you with a text editor. If you make an account, you can save your projects to access/download locally later on which is nifty. It's not ideal for making a full-scale app/long-term project but it's so useful for testing/converting something in a new/different language or you're just inebriated at the bars and teaching strangers/new friends "how it be" and "what it do" in the world of software development.

TL;DR version:

replit.com = Just a great idea that's done right and so ferda; 10 thumbs up out of 2